requirements.txt
Django==4.0.0
mysql-connector-python==8.0.26
python-dotenv==0.19.0
sendgrid==6.9.1
requests==2.26.0
firebase-admin==5.0.0
bcrypt==3.2.0
